% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_data.R
\name{diagnose_cross_sectional}
\alias{diagnose_cross_sectional}
\title{Diagnose Cross-sectional}
\usage{
diagnose_cross_sectional(df, useX = TRUE)
}
\arguments{
\item{df}{cross-sectional dataset}

\item{useX}{boolean, specifies whether baseline covariate should be used in
function determining diagnosis probability}
}
\value{

}
\description{
Diagnose Cross-sectional
}
