% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_data.R
\name{diagnose_cs_with_doctors}
\alias{diagnose_cs_with_doctors}
\title{Diagnose Cross-sectional dataset with doctor variation}
\usage{
diagnose_cs_with_doctors(df)
}
\arguments{
\item{df}{cross-sectional dataset generated by \code{\link{generate_cs_with_doctors}}}
}
\value{
diagnosed cross_sectional data set.
}
\description{
Add diagnosis labels (0 or 1) to each individual in a cross-sectional dataset
}
