% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_data.R
\name{diagnose_longitudinal}
\alias{diagnose_longitudinal}
\title{Diagnose Longitudinal}
\usage{
diagnose_longitudinal(df, useX = TRUE)
}
\arguments{
\item{df}{longitudinal dataset}

\item{useX}{boolean, specifies whether baseline covariate should be used in
function determining diagnosis probability}
}
\value{

}
\description{
Diagnose Longitudinal
}
