% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_data.R
\name{diagnosis_fn}
\alias{diagnosis_fn}
\title{Calculate Diagnosis Probability, depending on x}
\usage{
diagnosis_fn(severity, x, theta = c(-10, 20, -5))
}
\arguments{
\item{severity}{disease severity (in [0,1])}

\item{x}{binary covariate}

\item{theta}{numeric vector giving parameters of diagnostic function}
}
\value{

}
\description{
Calculate Diagnosis Probability, depending on x
}
