% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_data.R
\name{diagnosis_fn_by_betas}
\alias{diagnosis_fn_by_betas}
\title{Calculate Diagnosis Probability}
\usage{
diagnosis_fn_by_betas(severity, x, b0 = -10, bS = 20, bX = -5)
}
\arguments{
\item{severity}{disease severity (in [0,1])}

\item{x}{binary covariate}

\item{b0}{beta coefficient for intercept of diagnostic function}

\item{bS}{beta coefficient for severity in diagnostic function}

\item{bX}{beta coefficient for X in diagnostic function}
}
\value{
diagnosis probability between 0 and 1.
}
\description{
Calculate Diagnosis Probability
}
