% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_data.R
\name{diagnosis_fn_noX}
\alias{diagnosis_fn_noX}
\title{Calculate Diagnosis Probability, depending on x}
\usage{
diagnosis_fn_noX(severity)
}
\arguments{
\item{severity}{disease severity (in [0,1])}

\item{x}{binary covariate}
}
\value{

}
\description{
Calculate Diagnosis Probability, depending on x
}
