% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_data.R
\name{diagnosis_fn_X}
\alias{diagnosis_fn_X}
\title{Calculate Diagnosis Probability, depending on x}
\usage{
diagnosis_fn_X(x, severity)
}
\arguments{
\item{x}{binary covariate}

\item{severity}{disease severity (in [0,1])}
}
\value{

}
\description{
Calculate Diagnosis Probability, depending on x
}
