% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_cross_sectional}
\alias{generate_cross_sectional}
\title{Generate Data}
\usage{
generate_cross_sectional(n = 10000, prevalence = 0.15)
}
\arguments{
\item{n}{sample size}

\item{prevalence}{disease prevalence}
}
\value{
data frame of n individuals baseline binary covariate x and desease severity
}
\description{
Generate Data
}
