% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_cs_case4}
\alias{generate_cs_case4}
\title{Generate Cross Sectional Data with variation in presentation between groups}
\usage{
generate_cs_case4(n = 10000, prevalence = 0.1)
}
\arguments{
\item{n}{sample size}

\item{prevalence}{disease prevalence. Can be a single number from 0 to 1 or a
vector of two numbers.  If a vector, the first number will be the
prevalence in the group with x = 0 (usually the classical group), the second will be the prevalence in
the group with x = 1 (ususally the non-classical group).}
}
\value{
data frame of n individuals with baseline binary covariate x, disease
  status, and continuous disease severity
}
\description{
Produces a model data frame for a cross-sectional study of n individuals.
Allow non-diseased people to still have some symptoms, following a beta
distribution.  Individuals with covariate X = 1 have a different distribution
of symptoms than those with X = 0
}
