% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_cs_with_doctors}
\alias{generate_cs_with_doctors}
\title{Generate Cross Sectional Data with Doctors}
\usage{
generate_cs_with_doctors(
  patients_per_doc,
  n_docs,
  prevalence = 0.15,
  theta_mean = c(-10, 20, -5),
  sigma = 0.2
)
}
\arguments{
\item{patients_per_doc}{number of patients seen by each doctor}

\item{n_docs}{number of unique doctors in the data set}

\item{prevalence}{disease prevalence (for now, only supports one number for
prevalence - i.e. groups must have the same disease prevalence)}

\item{theta_mean}{mean prescribing practices}
}
\value{

}
\description{
Produces a model data frame for a cross-sectional study of individuals, each
assigned to a specific doctor with diagnostic practices described by beta0,
betaS, and betaX
}
