% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_longitudinal}
\alias{generate_longitudinal}
\title{Generate Longitudinal}
\usage{
generate_longitudinal(n = 10000)
}
\arguments{
\item{n}{sample size}
}
\value{

}
\description{
Generate Longitudinal
}
