% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{severity_fn}
\alias{severity_fn}
\title{Severity Function}
\usage{
severity_fn(t, beta = 0.5, T_shift = 50)
}
\arguments{
\item{t}{time (years)}

\item{beta}{slope of disease progression}

\item{T_shift}{shift for time of progression onset}
}
\value{
severity between 0 and 1
}
\description{
Calculate severity over time for longitudinal data based on an individual's
starting parameters.  Helper function for extend_longitudinal.
}
