% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_data.R
\name{test_longitudinal}
\alias{test_longitudinal}
\title{test Longitudinal Data Set}
\usage{
test_longitudinal(df, theta = c(-12, 20, -3))
}
\arguments{
\item{df}{longitudinal dataset}

\item{theta, }{specification for diagnosis function.  If "truth", p_diagnose
is the true probability that the individual has the disease given X and R}
}
\value{
diagnosed longitudinal data set.
}
\description{
At each time point for each individual, calculate probability of test
using diagnosis_fn.  If that individual is not yet diagnosed, check if a new
diagnosis would happen at this time point based on the diagnosis probability.
}
