% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_data.R
\name{diagnose_cross_sectional}
\alias{diagnose_cross_sectional}
\title{Diagnose Cross-sectional dataset}
\usage{
diagnose_cross_sectional(df, theta = c(-12, 20, -3))
}
\arguments{
\item{df}{cross-sectional dataset}

\item{theta, }{parameters for diagnosis function}
}
\value{
diagnosed cross_sectional data set.
}
\description{
Add diagnosis labels (0 or 1) to each individual in a cross-sectional dataset
}
