% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_data.R
\name{diagnose_longitudinal}
\alias{diagnose_longitudinal}
\title{Diagnose Longitudinal Data Set}
\usage{
diagnose_longitudinal(df, theta = c(-10, 20, -5))
}
\arguments{
\item{df}{longitudinal dataset}

\item{theta}{parameters for diagnosis function}
}
\value{
diagnosed longitudinal data set.
}
\description{
At each time point for each individual, calculate probability of diagnosis
using diagnosis_fn.  If that individual is not yet diagnosed, check if a new
diagnosis would happen at this time point based on the diagnosis probability.
}
