% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_data.R
\name{diagnosis_fn}
\alias{diagnosis_fn}
\title{Calculate Diagnosis Probability using theta}
\usage{
diagnosis_fn(representativeness, x, theta = c(-12, 20, -3))
}
\arguments{
\item{representativeness}{disease representativeness (in [0,1])}

\item{x}{binary covariate}

\item{theta, }{specification for diagnosis function.  If "truth", p_diagnose
is the true probability that the individual has the disease given X and R}
}
\value{
diagnosis probability between 0 and 1.
}
\description{
Calculate Diagnosis Probability using theta
}
