% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_longitudinal}
\alias{generate_longitudinal}
\title{Generate Longitudinal Data Set}
\usage{
generate_longitudinal(n = 10000, sigma = 2)
}
\arguments{
\item{n}{sample size}

\item{sigma}{noise to use when generating representativeness from severity}
}
\value{
a data.frame in long-form
}
\description{
Generate a long-form longitudinal data set of n individuals, each observed at
100 time points.  Each individual has the underlying disease, with a disease
severity that increases over time based on a logistic function.  The logistic
function is parametrized by beta (rate of onset) and T_shift (time of onset),
which varies between individuals.
}
