% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_data.R
\name{test_cross_sectional}
\alias{test_cross_sectional}
\title{Test Cross-sectional dataset}
\usage{
test_cross_sectional(df, theta = c(-10, 20, -5))
}
\arguments{
\item{df}{cross-sectional dataset}

\item{theta, }{parameters for diagnosis function}
}
\value{
diagnosed cross_sectional data set.
}
\description{
Add testing and diagnosis labels (0 or 1) to each individual in a
cross-sectional dataset
}
